Prepare to support multiple concurrent serial port connections
