+++ App.config
@@ -1,38 +1,38 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
     <configSections>
-        <sectionGroup name="applicationSettings" type="System.Configuration.ApplicationSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" >
+        <sectionGroup name="applicationSettings" type="System.Configuration.ApplicationSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
             <section name="IMMULIS.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
         </sectionGroup>
     </configSections>
+    <connectionStrings>
+        <add name="IMMULIS.Properties.Settings.ConnectionString" connectionString="Data Source=127.0.0.1\SQLExpress;Initial Catalog=LISDB;Integrated Security=True" providerName="" />
+    </connectionStrings>
     <startup> 
         <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.8" />
     </startup>
     <applicationSettings>
         <IMMULIS.Properties.Settings>
-            <setting name="SerialPortNum" serializeAs="String">
-                <value>COM4</value>
+            <setting name="SerialPorts" serializeAs="String">
+                <value>COM4,9600,None,8,1,None;COM5,19200,None,8,1</value>
             </setting>
             <setting name="SerialPortBaudRate" serializeAs="String">
                 <value>9600</value>
             </setting>
             <setting name="SerialPortParity" serializeAs="String">
-                <value>0</value>
+                <value>None</value>
             </setting>
             <setting name="SerialPortDataBits" serializeAs="String">
                 <value>8</value>
             </setting>
             <setting name="SerialPortStopBits" serializeAs="String">
-                <value>1</value>
+                <value>One</value>
             </setting>
             <setting name="LIS_ID" serializeAs="String">
                 <value>LIS</value>
             </setting>
             <setting name="LIS_Password" serializeAs="String">
                 <value />
-            </setting>
-            <setting name="ConnectionString" serializeAs="String">
-                <value>Data Source=127.0.0.1\SQLExpress;Initial Catalog=LISDB;Integrated Security=True</value>
             </setting>
             <setting name="SenderAddress" serializeAs="String">
                 <value />
@@ -49,6 +49,21 @@
             <setting name="AutoSendInterval" serializeAs="String">
                 <value>15000</value>
             </setting>
+            <setting name="SerialPortHandshake" serializeAs="String">
+                <value>None</value>
+            </setting>
         </IMMULIS.Properties.Settings>
     </applicationSettings>
+  <runtime>
+    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
+      <dependentAssembly>
+        <assemblyIdentity name="SQLitePCLRaw.core" publicKeyToken="1488e028ca7ab535" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-2.0.5.1333" newVersion="2.0.5.1333" />
+      </dependentAssembly>
+      <dependentAssembly>
+        <assemblyIdentity name="System.Runtime.CompilerServices.Unsafe" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-5.0.0.0" newVersion="5.0.0.0" />
+      </dependentAssembly>
+    </assemblyBinding>
+  </runtime>
 </configuration>
