+++ Constants.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace IMMULIS
+{
+     public static class Constants
+     {
+          public const string ACK = "\x06";
+
+          public const string CR = "\x0D";
+
+          public const string ENQ = "\x05";
+
+          public const string EOT = "\x04";
+
+          public const string ETB = "\x17";
+
+          public const string ETX = "\x03";
+
+          public const string LF = "\x0A";
+
+          public const string NAK = "\x15";
+
+          public const string STX = "\x02";
+     }
+}
