+++ IMMULIS.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 16.0.30804.86
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IMMULIS", "IMMULIS.csproj", "{2804EF32-781C-434A-9443-A2DCFF9CB60C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LISUnitTests", "..\LISUnitTests\LISUnitTests.csproj", "{3413CCC0-1AD1-4984-B550-F91F34506B57}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{2804EF32-781C-434A-9443-A2DCFF9CB60C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2804EF32-781C-434A-9443-A2DCFF9CB60C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2804EF32-781C-434A-9443-A2DCFF9CB60C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3413CCC0-1AD1-4984-B550-F91F34506B57}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3413CCC0-1AD1-4984-B550-F91F34506B57}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3413CCC0-1AD1-4984-B550-F91F34506B57}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3413CCC0-1AD1-4984-B550-F91F34506B57}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
