+++ Program.cs
@@ -11,7 +11,7 @@ namespace IMMULIS
           {
                if (System.Environment.UserInteractive)
                {    // Execute the program as a console app for debugging purposes.
-                    IMMULIService service1 = new IMMULIService();
+                    ServiceMain service1 = new ServiceMain();
                     service1.DebuggingRoutine(args);
                }
                else
@@ -20,7 +20,7 @@ namespace IMMULIS
                     ServiceBase[] ServicesToRun;
                     ServicesToRun = new ServiceBase[]
                     {
-                     new IMMULIService()
+                     new ServiceMain()
                     };
                     ServiceBase.Run(ServicesToRun);
                }
