+++ Properties/Settings.settings
@@ -2,20 +2,20 @@
 <SettingsFile xmlns="http://schemas.microsoft.com/VisualStudio/2004/01/settings" CurrentProfile="(Default)" GeneratedClassNamespace="IMMULIS.Properties" GeneratedClassName="Settings">
   <Profiles />
   <Settings>
-    <Setting Name="SerialPortNum" Type="System.String" Scope="Application">
-      <Value Profile="(Default)">COM4</Value>
+    <Setting Name="SerialPorts" Type="System.String" Scope="Application">
+      <Value Profile="(Default)">COM4,9600,None,8,1,None;COM5,19200,None,8,1</Value>
     </Setting>
     <Setting Name="SerialPortBaudRate" Type="System.Int32" Scope="Application">
       <Value Profile="(Default)">9600</Value>
     </Setting>
-    <Setting Name="SerialPortParity" Type="System.Int32" Scope="Application">
-      <Value Profile="(Default)">0</Value>
+    <Setting Name="SerialPortParity" Type="System.IO.Ports.Parity" Scope="Application">
+      <Value Profile="(Default)">None</Value>
     </Setting>
     <Setting Name="SerialPortDataBits" Type="System.Int32" Scope="Application">
       <Value Profile="(Default)">8</Value>
     </Setting>
-    <Setting Name="SerialPortStopBits" Type="System.Int32" Scope="Application">
-      <Value Profile="(Default)">1</Value>
+    <Setting Name="SerialPortStopBits" Type="System.IO.Ports.StopBits" Scope="Application">
+      <Value Profile="(Default)">One</Value>
     </Setting>
     <Setting Name="LIS_ID" Type="System.String" Scope="Application">
       <Value Profile="(Default)">LIS</Value>
@@ -23,7 +23,12 @@
     <Setting Name="LIS_Password" Type="System.String" Scope="Application">
       <Value Profile="(Default)" />
     </Setting>
-    <Setting Name="ConnectionString" Type="System.String" Scope="Application">
+    <Setting Name="ConnectionString" Type="(Connection string)" Scope="Application">
+      <DesignTimeValue Profile="(Default)">&lt;?xml version="1.0" encoding="utf-16"?&gt;
+&lt;SerializableConnectionString xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema"&gt;
+  &lt;ConnectionString&gt;Data Source=127.0.0.1\SQLExpress;Initial Catalog=LISDB;Integrated Security=True&lt;/ConnectionString&gt;
+  &lt;ProviderName /&gt;
+&lt;/SerializableConnectionString&gt;</DesignTimeValue>
       <Value Profile="(Default)">Data Source=127.0.0.1\SQLExpress;Initial Catalog=LISDB;Integrated Security=True</Value>
     </Setting>
     <Setting Name="SenderAddress" Type="System.String" Scope="Application">
@@ -40,6 +45,9 @@
     </Setting>
     <Setting Name="AutoSendInterval" Type="System.Double" Scope="Application">
       <Value Profile="(Default)">15000</Value>
+    </Setting>
+    <Setting Name="SerialPortHandshake" Type="System.IO.Ports.Handshake" Scope="Application">
+      <Value Profile="(Default)">None</Value>
     </Setting>
   </Settings>
 </SettingsFile>
