+++ TransENQState.cs
@@ -1,9 +1,11 @@
-﻿using static IMMULIS.IMMULIService;
+﻿using static IMMULIS.ServiceMain;
 
 namespace IMMULIS
 {
      class TransENQState : ILISState
      {
+
+          protected internal CommFacilitator comm;
           public void RcvInput(string InputString)
           {
                switch (InputString)
@@ -28,14 +30,14 @@ namespace IMMULIS
           public void RcvACK()
           {
                // Send next frame.
-               CurrentMessage = OutboundMessageQueue.Dequeue();
-               CurrentMessage.PrepareToSend();
-               ComPort.Send(CurrentMessage.FrameList[CurrentFrameCounter]);
-               CurrentFrameCounter++;
+               comm.CurrentMessage = comm.OutboundMessageQueue.Dequeue();
+               comm.CurrentMessage.PrepareToSend();
+               comm.ComPort.Send(comm.CurrentMessage.FrameList[comm.CurrentFrameCounter]);
+               comm.CurrentFrameCounter++;
                // Reset the NAK count to 0.
-               numNAK = 0;
+               comm.numNAK = 0;
                // Reset the transaction timer to 15 seconds.
-               transTimer.Reset(15);
+               comm.transTimer.Reset(15);
           }
 
           public void RcvData(string InputString)
@@ -47,7 +49,7 @@ namespace IMMULIS
           public void RcvENQ()
           {
                // Set the contention timer to 20 seconds and return to idle state.
-               ContentTimer.Reset(20);
+               comm.ContentTimer.Reset(20);
           }
 
           public void RcvEOT()
@@ -66,7 +68,7 @@ namespace IMMULIS
           public void RcvNAK()
           {
                // If the instrument responds to our ENQ with NAK, it's busy. Back to Idle for 10 seconds.
-               BusyTimer.Reset(10);
+               comm.BusyTimer.Reset(10);
           }
 
           void ILISState.HaveData()
