+++ TransWaitState.cs
@@ -1,10 +1,12 @@
 ﻿using System;
-using static IMMULIS.IMMULIService;
+using static IMMULIS.ServiceMain;
 
 namespace IMMULIS
 {
      class TransWaitState : ILISState
      {
+
+          protected internal CommFacilitator comm;
           public void RcvInput(string InputString)
           {
                switch (InputString)
@@ -29,24 +31,24 @@ namespace IMMULIS
           public void RcvACK()
           {
 #if DEBUG
-               AppendToLog("CurrentMessage.FrameList.Count: " + CurrentMessage.FrameList.Count);
-               AppendToLog("CurrentFrameCounter: " + CurrentFrameCounter);
+               AppendToLog("CurrentMessage.FrameList.Count: " + comm.CurrentMessage.FrameList.Count);
+               AppendToLog("CurrentFrameCounter: " + comm.CurrentFrameCounter);
 #endif
                // If all frames have been sent, end the transmission.
-               if (CurrentMessage.FrameList.Count == CurrentFrameCounter)
+               if (comm.CurrentMessage.FrameList.Count == comm.CurrentFrameCounter)
                {
-                    ComPort.Send(Constants.EOT);
-                    CurrentMessage = new MessageBody();
+                    comm.ComPort.Send(Constants.EOT);
+                    comm.CurrentMessage = new Message();
                }
                else
                {
                     // Otherwise, send next frame.
-                    ComPort.Send(CurrentMessage.FrameList[CurrentFrameCounter]);
-                    CurrentFrameCounter++;
+                    comm.ComPort.Send(comm.CurrentMessage.FrameList[comm.CurrentFrameCounter]);
+                    comm.CurrentFrameCounter++;
                     // Reset the NAK count to 0.
-                    numNAK = 0;
+                    comm.numNAK = 0;
                     // Reset the transaction timer to 15 seconds.
-                    transTimer.Reset(15);
+                    comm.transTimer.Reset(15);
                }
           }
 
@@ -55,7 +57,7 @@ namespace IMMULIS
                // Data frames should always be preceded by other signals, so for now, just log it.
                // Signal the instrument to interrupt the transmission with an EOT.
                AppendToLog("Data received in TransWait state: " + InputString);
-               ComPort.Send(Constants.EOT);
+               comm.ComPort.Send(Constants.EOT);
           }
 
           public void RcvENQ()
@@ -82,16 +84,16 @@ namespace IMMULIS
           public void RcvNAK()
           {
                // Send old frame.
-               ComPort.Send(CurrentMessage.FrameList[CurrentFrameCounter - 1]);
+               comm.ComPort.Send(comm.CurrentMessage.FrameList[comm.CurrentFrameCounter - 1]);
                // Increment NAK count.
-               numNAK++;
-               if (numNAK == 6)
+               comm.numNAK++;
+               if (comm.numNAK == 6)
                {
                     // Too many NAKs. Something's wrong. Send an EOT and go back to Idle.
                     // Maybe stick the message back in the queue to try again later?
-                    ComPort.Send(Constants.EOT);
-                    OutboundMessageQueue.Enqueue(CurrentMessage);
-                    CurrentMessage = new MessageBody();
+                    comm.ComPort.Send(Constants.EOT);
+                    comm.OutboundMessageQueue.Enqueue(comm.CurrentMessage);
+                    comm.CurrentMessage = new Message();
                }
           }
