+++ packages.config
@@ -1,0 +1,15 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.Data.Sqlite" version="5.0.9" targetFramework="net48" />
+  <package id="Microsoft.Data.Sqlite.Core" version="5.0.9" targetFramework="net48" />
+  <package id="SQLitePCLRaw.bundle_e_sqlite3" version="2.0.5" targetFramework="net48" />
+  <package id="SQLitePCLRaw.core" version="2.0.5" targetFramework="net48" />
+  <package id="SQLitePCLRaw.lib.e_sqlite3" version="2.0.5" targetFramework="net48" />
+  <package id="SQLitePCLRaw.provider.dynamic_cdecl" version="2.0.5" targetFramework="net48" />
+  <package id="SQLitePCLRaw.provider.e_sqlite3" version="2.0.5" targetFramework="net48" />
+  <package id="System.Buffers" version="4.5.1" targetFramework="net48" />
+  <package id="System.Memory" version="4.5.4" targetFramework="net48" />
+  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net48" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="5.0.0" targetFramework="net48" />
+  <package id="YamlDotNet" version="11.2.1" targetFramework="net48" />
+</packages>
